More work on helpers.
Allow passing multiple arguments to helpers:
any conf option prefixed with helper_ will be passed in
a helper_args dictionary.

nyt_selenium: allow specifying the log file location as well as
the executable path, and if no log file is specified, use tempdir
to create a reasonable filename (instead of geckodriver's option
of putting it in . regardless of whether . is actually writable).