helper argument improvements.
Add a $f option which is $d/feedname.
If $d or $f is specified in the selenium helper log_file,
create it before creating the webdriver.Firefox, otherwise
the webdriver will refuse to run.
Update the documentation for $d and $f.
Rename a few variables related to the feeds dir to be clearer.