Use img data-src tag in preference to src.
The LA Daily Post has once again changed the wacky way they
deal with images; images have a src="data:" tag that is completely
bogus and displays nothing, but probably signals to some JS that
the real src URL is in the data-src attribute.
So, to make the LADP work, use data-src in preference to src for
img tags, and hope that no one else is setting data-src for a
different purpose.