Don't overwrite feeds already fed.
If a site's feed dir for today exists and has an index.html,
assume that feed is done, and don't overwrite it.
This (hopefully) will eliminate the need to do the cache dir dance
when restarting after an error.
Alas, it doesn't help anyone coming in via FeedViewer, since FeedViewer
(currently) refuses to run if the site feed dir already exists.